_face_ab_test/requirements.txt
insightface>=0.7.3
deepface>=0.0.92
facenet-pytorch>=2.6.0
onnxruntime-gpu>=1.17.0
opencv-python-headless>=4.9.0
scikit-learn>=1.4.0
numpy>=1.26.0
Pillow>=10.0.0
ultralytics>=8.1.0
mediapipe>=0.10.0
